refactor(pumbility): records and events carry doubles
The four pools stop being Rating -- that value type is int-backed, so a
record holding one had already spent the precision before any caller
could decide to. PlayerStatsRecord, CommunityLeaderboardRecord,
CommunityPlayerProfileRecord, PlayerRatingRecord, HighlightDetail and
PlayerRatingsImprovedEvent all widen. TotalRating is not a pool -- it is
the lifetime sum under its own label -- so it stays Rating.

Two narrowings that were quietly costing accuracy go with them:

- PlayerRatingSaga.RankOn took an int, so a pool was seated against the
  official board at whole-number precision even though the board reads
  in decimals.
- SupplementRollupSaga.PlayerRatings held ints and then cast to decimal
  on the way into board rows, rounding immediately before widening again.

Formatting is untouched -- every site still renders N0 -- so nothing on
screen moves. Suites green: 3328 / 736 / 148.

+++ ScoreTracker/ScoreTracker.Domain/Events/PlayerRatingsImprovedEvent.cs
@@ -2,11 +2,15 @@
 
 namespace ScoreTracker.Domain.Events
 {
+    /// <summary>
+    ///     The top-50 pools and the co-op rating ride as doubles: they are PUMBILITY, and nothing
+    ///     below the presentation layer rounds it.
+    /// </summary>
     [ExcludeFromCodeCoverage]
-    public sealed record PlayerRatingsImprovedEvent(Guid UserId, int OldTop50, int OldSinglesTop50,
-        int OldDoublesTop50, int NewTop50, int NewSinglesTop50, int NewDoublesTop50, double OldCompetitive,
+    public sealed record PlayerRatingsImprovedEvent(Guid UserId, double OldTop50, double OldSinglesTop50,
+        double OldDoublesTop50, double NewTop50, double NewSinglesTop50, double NewDoublesTop50, double OldCompetitive,
         double NewCompetitive, double OldSinglesCompetitive, double NewSinglesCompetitive, double OldDoublesCompetitive,
-        double NewDoublesCompetitive, int CoOpRating, int PassCount, MixEnum Mix, Guid? SessionId = null)
+        double NewDoublesCompetitive, double CoOpRating, int PassCount, MixEnum Mix, Guid? SessionId = null)
     {
     }
 }
